Commit #2 Branch2
in the NV, from line 1-52, This LoginForm class is a Swing JFrame that lets a user log in with their username and password. It receives the list of registered User objects from the SignUp form, then checks each user to see if the credentials match. If a match is found, it opens the Dashboard; otherwise, it shows an “Invalid username or password” message.